test: retries on e2e matrix strategy and matrix generator jobs (#3009)
<!--
Thank you for your Pull Request! Please describe the problem this PR
fixes and a summary of the changes made.
Link to any relevant issues, code snippets, or other PRs.

For trivial changes, this template can be ignored in favor of a short
description of the changes.
-->

## Problem

<!--
Describe the issue this PR is solving
-->

E2E tests can fail for a variety of reasons, and matrices are especially
vulnerable to intermittent failure, due to the aggregation of several
parameterized runs.

## Changes

<!--
Summarize the changes introduced in this PR. This is a good place to
call out critical or potentially problematic parts of the change.
-->

Implement retries for steps for which we anticipate or commonly see
failures.

## Validation

<!--
Describe how changes in this PR have been validated. This may include
added or updated unit, integration and/or E2E tests, test workflow runs,
or manual verification. If manual verification is the only way changes
in this PR have been validated, you will need to write some automated
tests before this PR is ready to merge.

For changes to test infra, or non-functional changes, tests are not
always required. Instead, you should call out _why_ you think tests are
not required here.

If changes affect a GitHub workflow that is not included in the PR
checks, include a link to a passing test run of the modified workflow.
--->

Run checks that include E2E tests.

## Checklist

<!--
These items must be completed before a PR is ready to be merged.
Feel free to publish a draft PR before these items are complete.
-->

- [x] If this PR includes a functional change to the runtime behavior of
the code, I have added or updated automated test coverage for this
change.
- [x] If this PR requires a change to the [Project Architecture
README](../PROJECT_ARCHITECTURE.md), I have included that update in this
PR.
- [x] If this PR requires a docs update, I have linked to that docs PR
above.
- [x] If this PR modifies E2E tests, makes changes to resource
provisioning, or makes SDK calls, I have run the PR checks with the
`run-e2e` label set.

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license._