chore: increase retry waiting intervals and attempt counts (#3018)
<!--
Thank you for your Pull Request! Please describe the problem this PR
fixes and a summary of the changes made.
Link to any relevant issues, code snippets, or other PRs.

For trivial changes, this template can be ignored in favor of a short
description of the changes.
-->

## Problem

<!--
Describe the issue this PR is solving
-->

Even though we already retry e2e jobs 3 times, we are still getting
failures on the scheduled `health_checks` workflow.

ex.
https://github.com/aws-amplify/amplify-backend/actions/runs/18258502550/job/51983262437

## Changes

<!--
Summarize the changes introduced in this PR. This is a good place to
call out critical or potentially problematic parts of the change.
-->

Increased the number of attempts to 5. Increased the waiting interval
between retries from 30 seconds to 300 seconds.

## Validation

<!--
Describe how changes in this PR have been validated. This may include
added or updated unit, integration and/or E2E tests, test workflow runs,
or manual verification. If manual verification is the only way changes
in this PR have been validated, you will need to write some automated
tests before this PR is ready to merge.

For changes to test infra, or non-functional changes, tests are not
always required. Instead, you should call out _why_ you think tests are
not required here.

If changes affect a GitHub workflow that is not included in the PR
checks, include a link to a passing test run of the modified workflow.
--->

This PR will be run with the workflow, with the `run-e2e` label applied.

## Checklist

<!--
These items must be completed before a PR is ready to be merged.
Feel free to publish a draft PR before these items are complete.
-->

- [X] If this PR includes a functional change to the runtime behavior of
the code, I have added or updated automated test coverage for this
change.
- [X] If this PR requires a change to the [Project Architecture
README](../PROJECT_ARCHITECTURE.md), I have included that update in this
PR.
- [X] If this PR requires a docs update, I have linked to that docs PR
above.
- [X] If this PR modifies E2E tests, makes changes to resource
provisioning, or makes SDK calls, I have run the PR checks with the
`run-e2e` label set.

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license._