fix: workflows cancelling on wrong scope of timeout (#3023)
<!--
Thank you for your Pull Request! Please describe the problem this PR
fixes and a summary of the changes made.
Link to any relevant issues, code snippets, or other PRs.

For trivial changes, this template can be ignored in favor of a short
description of the changes.
-->

## Problem

<!--
Describe the issue this PR is solving
-->

<img width="1439" height="755" alt="Screenshot 2025-10-14 at 11 14 03"
src="https://github.com/user-attachments/assets/ce006e1b-3cbc-4064-bd5c-254f28e46c59"
/>
Some workflows have timeouts that are too short to account for retries
in GitHub action `run_with_e2e_account`. That is causing test runs with
retries to occasionally cancel themselves early.

## Changes

<!--
Summarize the changes introduced in this PR. This is a good place to
call out critical or potentially problematic parts of the change.
-->
Implemented per-attempt timeouts within the `e2e` runner and increased
overall timeouts.

Acknowledged risk: since we increase the total timeout of the e2e runs -
if a task within the job hangs (and isn't the job being retried), it
will hang for the rest of the timeout. This was true before as well, but
increasing this number exacerbates it.

## Validation

<!--
Describe how changes in this PR have been validated. This may include
added or updated unit, integration and/or E2E tests, test workflow runs,
or manual verification. If manual verification is the only way changes
in this PR have been validated, you will need to write some automated
tests before this PR is ready to merge.

For changes to test infra, or non-functional changes, tests are not
always required. Instead, you should call out _why_ you think tests are
not required here.

If changes affect a GitHub workflow that is not included in the PR
checks, include a link to a passing test run of the modified workflow.
--->

The workflow run (with the `run-e2e` label) on this PR will validate
that the change doesn't break anything, but validation that it works in
the long term remains to be seen.

## Checklist

<!--
These items must be completed before a PR is ready to be merged.
Feel free to publish a draft PR before these items are complete.
-->

- [X] If this PR includes a functional change to the runtime behavior of
the code, I have added or updated automated test coverage for this
change.
- [X] If this PR requires a change to the [Project Architecture
README](../PROJECT_ARCHITECTURE.md), I have included that update in this
PR.
- [X] If this PR requires a docs update, I have linked to that docs PR
above.
- [X] If this PR modifies E2E tests, makes changes to resource
provisioning, or makes SDK calls, I have run the PR checks with the
`run-e2e` label set.

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license._