chore: bump SDK dependencies
In the health checks we're seeing the following SDK error:

```
Error: serializerMiddleware is not found when adding endpointV2Middleware middleware before serializerMiddleware
```

This seems to indicate incompatible versions somewhere in
the SDK/Smithy stack of dependencies.

Try to upgrade them all to a consistent number to see if that fixes it.