chore: bump SDK dependencies (#3057)
In the health checks we're seeing the following SDK error:

```
Error: serializerMiddleware is not found when adding endpointV2Middleware middleware before serializerMiddleware
```

This seems to indicate incompatible versions somewhere in the SDK/Smithy
stack of dependencies.

Try to upgrade them all to a consistent number to see if that fixes it.

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license._