chore: remove dependency on node built-in (#3058)
The health checks are now failing with

```
error TS2307: Cannot find module 'node:crypto' or its corresponding type declarations.
```

What the test does is `cdk init` a project, then remove the
`node_modules` directory in order to source the node modules from
elsewhere.

Rather than try to figure out where the package is installed and what
changed where in order to get the right definition of `@types/node`
anywhere, I figured it would be easier to just cut the single dependency
on the node built-in, and make the test sliiiightly more robust against
what types are installed.


_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license._